Make the --scripts-version option work with tests.
Using react-scripts-ts as the scripts version results in the test script
failing due to "--scripts-version" being an unsupported option for the
react-scripts-ts test script.

Since that option is only of relevance to react-app-rewired, strip it out
of the list of arguments before passing those on to the called test script.