Fix issue 136.
The current way of searching for the oneOf rule breaks for anyone using the custom-scripts options if the scripts they are using do not have the exact layout that react-scripts does.

For example, react-scripts-ts has pre rules for both ts and for js before defining the oneOf rules, which means that it is in location rules[2] not rules[1].

This change searches for an actual rule defining the oneOf condition.