Document optional `override` option for eslint (#216)
As of #122, react-app-rewire-eslint takes an optional third argument, a function that allows further customization of the eslint options that are passed on to `eslint-loader`.  However, that function isn't documented, so people might not be aware that it exists.

This adds documentation and example code to show the use of this new argument.