Update readme, get Travis tests working, & exclude the /test app directory from npm packages. (#371)
* Update readme to cover version requirements.

Covers common issues such as the need to use react-app-rewired 1.x for
customised scripts versions based on create-react-app 1.x, and that the
eject script should not be rewired.

* Add required setup/teardown scripts to travis build

* Don't include the /test directory in npm packages.

* Minimum node version for CRA 2.x is 8.10.0.

Dependencies of the test app require a minimum node version of 8.10.0,
so testing that app with Node 6 will never succeed. Change test to use
Node 8 & 10.